backend/requirements.txt
asgiref==3.2.10
astroid==2.4.1
beautifulsoup4==4.9.1
bootstrap4==0.1.0
Django==3.0
django-bootstrap4==2.1.1
django-filter==2.3.0
djangorestframework==3.11.0
importlib-metadata==1.6.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pylint==2.5.2
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0
requests=2.24.0
